[regexp] Fix nested duplicate named capture groups
If named captures with the same name were nested within the same
disjunct, they were previously wrongly accepted.
This was due to all of the nested named captures in the first disjunct
being added to the non-participating named capture groups when parsing
the '|' to start a new alternative in the outermost group.
When the outermost group is closed and the capture name of the group is
checked, the nested group names were part of the non-participating
group, thus allowing duplicates within nested groups.
To avoid this, we explicitly check for duplicates in nested groups.

Fixed: 437825937
Change-Id: I4975259df36b77863a6d0d2fd4fe98f3e15dacde
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6859423
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101931}