[heap] Remove initialization fence on merging compacton spaces
On merging compaction spaces the pages are already fully
initialized.

This change affects only TSAN builds.

The call is a problem on TSAN builds as it atomically sets the
metadata pointer which then leads to reporting data races with
any access on the metadata object.

Bug: 429538831
Change-Id: I830f03279f7e702a429f63e22b0bd8fed055bdfb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6867872
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101964}