[d8] Expect limit exceeded in string concatenation
Method v8::String::Concat does not throw an exception in case the
length of the result exceeds the allowed limit. Instead, it returns
an empty local handle. This CL fixes a number of calls of this method
in d8, where the result can get arbitrarily big and this case was not
properly checked.

Bug: 437956096
Change-Id: Ie9f7127594fd1669b2e789340305ed2675d1ad3d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6863276
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101968}