[typed-array] Read length & elements kind once for copy
When copying from one TypedArray to another, read off the elements kind
only once, and calculate type and element size from that one read. Also,
only ever use the length that was passed in, to avoid multiple reads of
the length -- calculate the byte length locally from the given length,
rather than by re-reading the length.

Additionally, clean up handling of the destination type -- this should
already be handled by the templating of the helper class being used, so
we can define the destination type and size as constexpr (and just
DCHECK that it matches the input).

Finaly drive-by, remove the element type parameter from the helper
classes, and infer it from the ElementsKind, so that it's known to
always be consistent.

Fixed: 436140742
Change-Id: I930cfdb06f524602b256d0ce2d8692b4f0ead343
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6871394
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101990}