[wasm][rab] Wasm Memory RAB integration fixup
This addresses the issue discussed at:
https://github.com/WebAssembly/spec/issues/1895

Summary: the spec allows the {maxByteLength} property of a wasm memory
buffer to be up to 2^32 (= 2^16 maximum pages * 2^16 bytes per page),
which cannot be represented in the size_t {max_byte_length} field of the
buffer on 32 bit. A similar issue happens on 64 bit with memory64. The
solution agreed upon in the GitHub thread is to ignore the
{max_byte_length} field in the {maxByteLength} getter and recompute the
value instead.

There are other references to the {max_byte_length} field inside V8,
therefore we still compute the expected value in
{WasmMemoryObject::FixUpResizableArrayBuffer} and store it in the field.
And for the exceptional case where it is not representable, we keep the
old hack that consists in subtracting one page size from the value. But
this hack is no longer expected to be observable.

Another issue discussed in the thread is that with memory64,
{maxByteLength} is potentially bigger than MAX_SAFE_INTEGER (= 2^53-1).
However, as discussed in the thread, there is an implementation-defined
limit of 2^18 for memory64's maximum pages [1], and in V8 this limit is
validated when we create the memory (which is not the case for all
engines, there is an open issue about that here [2]). With our
current interpretation of the limit, {maxByteLength} is at most 2^18 *
2*16, which is less than MAX_SAFE_INTEGER.

[1] https://webassembly.github.io/memory64/js-api/#limits
[2] https://github.com/WebAssembly/spec/issues/1863

R=jkummerow@chromium.org

Bug: 42202693
Change-Id: I55dfb2acc6ddb036ec9b40c33a978d222f06bae3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6871493
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101986}