[heap] Add WriteBarrierModeScope
This CL adds a new class WriteBarrierModeScope which we can use
in debug builds to store the object the WriteBarrierMode was
computed for by GetWriteBarrierModeForObject. We can use this
additional information to strengthen the write barrier verification.

In order to do so, this CL adds an additional WriteBarrierMode
SKIP_WRITE_BARRIER_SCOPE. Any time this mode is used, the host
object of the write barrier has to match the object stored in
WriteBarrierModeScope.

The split between SKIP_WRITE_BARRIER and SKIP_WRITE_BARRIER_SCOPED
will also enable a follow-up CL to strength the write barrier
verification where write barriers are removed statically.

Bug: 437096305
Change-Id: Iaba6472e49c4ea34884f0707afd3309b984eaa54
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6839044
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102006}