[json] Parser: Fix fast-iterable key detection
Checking if a potentially fast-iterable key matches the key in the
feedback was broken since a previous fix [1].
This "fix" introduced scanning of the property key prior to checking if
it's content matches the expected key to check other properties, but
scanning the property key advanced the cursor and FastKeyMatch compared
the contents at the current position.
Fix this by checking the contents of the already scanned key against the
expected key value.

[1] https://crrev.com/c/6511763

Fixed: 440279755
Change-Id: Iab1d60e18eb8c0a7bc0c15cd967883f0c0882d06
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6879445
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Auto-Submit: Patrick Thier <pthier@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102014}