[logging] Append IDLE_EXTERNAL to StateTag to preserve log compatibility
A new VM state `IDLE_EXTERNAL` was added in https://crrev.com/c/6714365,
but inserted in the middle of `StateTag` enum instead of appended
to the end. Because the numeric value of `StateTag` is written into tick
events in V8 log and used to attribute time, inserting a value mid-enum
shifted all subsequent state codes (LOGGING changed from 9->10).

This CL moves `IDLE_EXTERNAL` to the end of the enum to restore previous
numeric codes for existing states. This is a minimal change that
re-establishes compatibility for current and historical logs.

Bug: 440481920
Change-Id: I2afa73f5140fd87f74721af9312f5fe7b29aac51
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6875647
Commit-Queue: Roman Dvornov <romandvornov@microsoft.com>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102026}