Add checks after GetIsolateFromHeapObject
This refactors one call site to just use `Isolate::Current()`, and adds
`DCHECK`s for others to verify that the result is identical to
`Isolate::Current()`.

If this CL sticks for a while, we can replace those calls by just
`Isolate::Current()` for improved performance and security.

R=leszeks@chromium.org

Bug: 431584880
Change-Id: Iba98d70aafee995b0ac81407c6235dd739603567
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6884670
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102057}