[wasmfx] Add a simple continuation wrapper
The purpose of this wrapper is to adapt the 'resume' call interface
to the func ref's call interface when we resume a continuation for
the first time. It will also be responsible for switching back to
the parent stack on return, and forwarding the return values.

The current version is very incomplete and is essentially a
placeholder: it does not handle arguments, return values, and
does not switch stacks yet. The change focuses on supporting the
new wrapper kind, compiling it, linking it and calling it.

- Compile the continuation wrapper during module compilation and store
  it in the NativeModule.
- During cont.new, initialize the jump buffer's PC to the continuation
  wrapper.
- In resume, load the PC from the target stack and call it. Pass a
  pointer to the resumed StackMemory as an argument.
- In the wrapper, load the func ref from the StackMemory argument and
  call it.
- Add new enum values to represent the new code kind, frame type, and
  call kind.

Bug: 388533754
Change-Id: I3fc0fab044570b5a69f36ff76d417a9c73d6ef68
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6827264
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102062}