[fuzzer][wasm] fix fuzzer performance drop due to comparing big linear memories
The issue is that when getting the byte for the minimum number of pages,
often, there are no more data bytes available in the DataRange, meaning
it returns 0. This leads to generating large linear memories in a lof of
cases, instead of the intended 1/256 chance. This slows down the fuzzing
because of the differential fuzzing comparing the memories between the
reference run and the actual run.

Bug: 436219201
Change-Id: I64d895383d7a9c2229752d8ae20fa63498d5e498
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6890352
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Nikolas Kaipel <nikolaskaipel@google.com>
Cr-Commit-Position: refs/heads/main@{#102097}