[heap] Rework COMPACTION_WAS_ABORTED
This moves the flag that signals that compaction was aborted to the
metadata object.

The tricky part here is that when aborting compaction due to OOM we
may leave behind multiple half-evacuated pages that we need to
re-record pointers on. Such pages must:
- still be evacuation candidates as otherwise they won't trigger
  recording;
- aborted and to signal that we should not skip recorded,
  as we would otherwise to on evacuation candidates;

Since the check for aborted evacuation candidates now becomes more
expensive, only the visitor that implements recording of slots during
the atomic pause (RecordMigratedSlotVisitor) implements this
functionality.

The rest of the CL is merely refactoring the calls a bit.

Change-Id: I1ef717b11c0566642bf5c58603c9b255a6d77d90
Bug: 429538831
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6884745
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102095}