[test] Derive FUZZTEST_PRNG_SEED from random seed
Instead of using a fixed seed, derive the seed from the runner's random
seed. This has multiple advantages:
- Pre-existing bugs in fuzzers to not block CLs which change the fuzzer
  and accidentally uncover the bug (there might be flakes though).
- The CQ and the waterfall provide some fuzzing coverage by running each
  fuzztest for 1 second on each run.
- You get a lightweight (but rather inefficient) fuzzer via
  `--random-seed-stress-count`.

The obvious disadvantage is that this might introduce (more) flakes.
Note that the current fixed seed already was not fully deterministic
because the number of executed test cases was not fixed.

In case of flakes, we can still use flako though, either with a fixed
random seed for more determinism, or without to bisect over CLs which
change the fuzzer's implementation (and it's generation or
interpretation of input).

R=machenbach@chromium.org
CC=mliedtke@chromium.org

Change-Id: I1755bec6c0b81fcf289e1187af4ea2fefc5d9baf
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6895698
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102108}