[regexp] Add ClearRegisters bytecode
Previously we translated ClearRegisters to a sequence of SetRegister
bytecodes. However SetRegister is not meant to be used for capture
registers. The interpreter ignored this limitation so far.
Besides fixing a DCHECK when we assemble RegExp code from bytecode, we
can also save on bytecode size if multiple sequential registers need to
be cleared.

Bug: 437003349
Change-Id: I8eebd57ce48f102fc993203913a90534971d8104
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6888831
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102125}