[maglev] Fix regressions when not eliding ReturnedValue
After non-eager inlining a function, we overwrite the call
node with a ReturnedValue or Identity node.

The deopt frame (from the inline call) could be dropped if
the inline function has no deopting node. If this deopt frame
points to a previous ReturnedValue, this would be retained
even if unused and not removed from the graph.

Crypto-md5-SP regresses because of that.

This CL refactors RemoveIdentities to be call in every
OverwriteNodeWith<Identity/ReturnedValue>.

Bug: 424157317
Change-Id: I384c632c008b26548e0269724ea781cc00fb6cb7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6898175
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102128}