[JS-Fuzzer] Add differential-fuzzing-specific corpora
This enables using the separate diff-fuzz corpora from Fuzzilli on
the differential-fuzzing instances of JS fuzzer. We now select
either differential-fuzzing or normal-fuzzing folders on respective
fuzzing instances. With a 20% chance we select the inverse.

To make this work, this change also refactors parts of the existing
code:
To pass through the information if we use differential fuzzing, we
parameter passing along the chain to construct a corpus. The control
flowroughly goes through:

run (entry file)
->script mutator (orchestrates fuzzing)
->runner (orchestrates composition of corpora)
->corpus

We refactor the way parameters are passed through this chain,
mainly adding various single parameters (e.g. number of files)
to a settings object.

The file selection is implemented using the corpus filter method.

Bug: 422361840
Change-Id: I541d91141f90adee690be45807f4e78cae33acad
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6818345
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102159}