[heap] Funnel read-only allocations into read-only region
For configuration that use pointer compression and relies on a shared
single data cage, which is the default in Chrome on 64-bit,  this CL
carves out an explicit contiguous memory region for read-only pages.

Additionally, we also make sure that the same region is not used by
the TrustedRange and CodeRange.  As a consequence we can check for
read-only objects by merely inspecting the object's address.

The change here does not remove the page flag yet but merely rewire
the allocations and use the condition in a DCHECK.

In a follow up this property can be used for:
- Checking for RO objects locally which is faster than checking page
  flags. It can also be combined with a partial Smi check (when we
  know that we are dealing with non-InstructionStream slots).
- Moving the RO page bit away from MemoryChunk.

A side effect is that this also filters out (more) small offsets for
conservative stack scanning as the contiguous region is larger than
the current static roots.

Change-Id: I3888c74a5d555cd78b81b5cf7c9f408edeac6053
Bug: 429538831
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6897249
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102168}