[wasm] Fix incorrect exception message in wasm trap handler
In the wasm runtime trap handler, the exception message depends on
whether the instruction that caused the trap is a shared-everything or a
linear memory atomic instruction.

This change fixes a bug where a linear memory atomic instruction can be
incorrectly identified as a shared-everything instruction if it is
encoded with more than one byte (excluding prefix), leading to an
incorrect exception message.

Fixed: 441816628
Change-Id: I1ff740912eaf46274c80ad55807bcf2f7c44e256
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6904555
Commit-Queue: Nikolas Kaipel <nikolaskaipel@google.com>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102208}