[wasmfx] Implement the resume instruction
Switch to the target stack in the resume instruction, and switch back to
the current parent stack on return from the stack entry wrapper.

Introduce the WasmFXResume and WasmFXReturn builtins to handle
the platform specific logic of saving/restoring the stack state. The
builtins reuse JSPI's {Isolate::SwitchStacks} via external C calls for
the shared platform-independent logic.

Add a new frame type for the WasmFXResume builtin to represent the stack
exit, and update the stack frame iterator to support iterating chains of
WasmFX stacks.

With this change, it is already possible for JSPI to attempt to suspend
multiple WasmFX stacks at once. However, properly supporting this
requires non-trivial changes and will be done in a separate change.

R=clemensb@chromium.org

Bug: 388533754
Change-Id: I5c96b0014ab3db3a6eb8472ec6bc346ee93a2e9f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6904557
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102193}