[heap] Rework lifetimes of MemoryChunk and MemoryChunkMetadata
MemoryChunk and MemoryChunkMetadata are acquired during page
allocation. When entering the MemoryPool we would previously keep the
objects alive in a possible broken state.

We would just keep the objects and allocate new ones over them when
acquiring memory from the MemoryPool. This somewhat worked as some
accessors on MemoryChunkMetadata were still valid. It breaks though
when trying to e.g. retrieve the owner() for a RO space check on the
metadata object as the previously owning Isolate may have already been
destroyed.

This CL now reworks this logic:
- MemoryChunk and MemoryChunkMetadata always use placement new() when
  initializing them for types that can enter the MemoryPool.
- Upon adding MemoryChunk and MemoryChunkMetadata to the pool we
  properly call destructors (which also already reset the
  chunk-to-metadata link) and create a value object that holds the
  memory reservations.
- When returning memory from the pool we use these uninitialized void*
  pointers for re-allocating the corresponding objects via placement
  new().
- Finally, we use trusted accessors when setting up the
  chunk-to-metadata link.

Drive-by: Fix IsMutablePageMetadata() only checking owner() field
which only works for detached RO heaps but misses the case when RO
heap is still writable.

Change-Id: Id8945b7fdc07e55b21d80a1d67df62c0d1c098fc
Bug: 429538831
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6903968
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102201}