[heap, test] Speculative fix for race in stack scanning test
This CL is a speculative fix for the race in the linked bug as I
was not able to reproduce it. I suspect that concurrent marking
here simply runs concurrently to the manual invocation of
IteratePointersForTesting(). Using the ManualGCScope should avoid
this by either finishing concurrent marking or by not starting it.

Bug: 442344808
Change-Id: I2d397244d32d4acd77d3b4e1fbc6e1ca987fa3c1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6909369
Auto-Submit: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Nikolaos Papaspyrou <nikolaos@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102210}