[turboshaft] Fix false-positive around strings in Load Elim verifier
This CL includes a revert of the previous "fix",
https://crrev.com/c/6829829, which tuned down load elimination to get
rid of the false positive. Now that we know that it was a false
positive, we can revert load elimination to its previous state and
just fix the verifier instead.

Fixed: 436623746
Change-Id: I38bae084d995985631291946f4ed60c1b00c99b7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6905293
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102216}