Add constness DCHECKs to WeakCell
finalization_registry_ and holdings_ are never updated after they are
initialized and could have been const. Keep the setter for the
initialization with a write barrier, but add a DCHECK that the fields
were uninitialized before. finalization_registry_ also is never
Undefined so it doesn't need to be a union of types.

target_ and unregister_token_ are const-like. They can be set once to
some object, and from then can only be re-set to Undefined.

WeakCell initialization in `finalization-registry.tq` is still bypassing
these c++ definitions by using field offset. This will be addressed
separately.

Bug: 340777103
Change-Id: I7fe9dd117fb69252275eaf4d34a84ca9d88d0aa7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6909453
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102233}