[test] Make bytecode expectation tests read golden snippets
Make the golden file snippets the source of truth for bytecode
expectations tests, to remove the duplicated snippets between the
bytecode-generator-unittest and the golden files. This should make it
simpler to add new expectation tests, by adding a snippet to a golden
file and rebaselining. It also unifies the handling of the header
options, like wrapping and flags.

The expectations parsing logic is extracted to a new
`BytecodeExpectationsParser`, which is used by both the test and the
expectations generator.

Change-Id: I6d82073233cf6abe268bf9137a862d215e4e27a0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6912523
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102236}