[maglev] Try to untag phis to Int32 instead of Float64 for int HeapNums
When a 32-bit integer doesn't fit in a Smi, V8 will box it to a
HeapNumber, and Maglev will then assume that it's a double.

For instance,

        function foo() {
          let i = 1732584193; // doesn't fit in a Smi

          for (; i == 1732584193;) {
            i = i ^ 1;
          }

          return i;
        }

        %PrepareFunctionForOptimization(foo);
        print(foo());
        print(foo());

        %OptimizeMaglevOnNextCall(foo);
        print(foo());


Before this CL, `i` would have been untagged to Float64, and it will
now be untagged to Int32 instead.

Bug: 431933185
Change-Id: I7cbcb30a7336f464e5dc23c8e6652ecfffede260
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6912787
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102243}