[wasm][jspi] Avoid gaps in the central stack
When we enter a WebAssembly.promising call (from the central stack), we
save the current SP in a global. If we need to reenter the central
stack for a JS/C call, we reenter at the saved address.

If WebAssembly.promising calls are nested, the innermost address
overwrites the previous one, and the previous value is lost. So at this
point, if we return (or throw or suspend) to the outer scope, and wasm
calls a JS/C function, it will use the lower address, leaving a gap in
the central stack.

Fix this by saving the previous address in the outer StackMemory object
and restoring it on return/suspend.

Drive-by: misc cleanups in Isolate::SwitchStacks().

Bug: 42202153
Change-Id: Ie6b5750a90a67757772d7e4f21f411dafeff31b2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6912743
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102331}