[wasm][turbolev] Wasm-in-JS inlining in Turboshaft
Inline the JS-to-Wasm wrappers in Turboshaft, to make JS-to-Wasm
calls more efficient.

This implementation works with Turbolev and reuses the
`WasmInJSInliningReducer`:

- In the Turbolev graph builder, when we process a `CallKnownJSFunction`
node for a call to a JS-to-Wasm wrapper, we add a `JSWasmCallParameters`
to the TSCallDescriptor.

- In the WasmInJSInliningReducer::REDUCE(Call), we try to inline the
wrapper.

- The actual inlining is done by
`WasmWrapperTSGraphBuilder::BuildJSToWasmWrapperImpl`. All the logic to
build a JS-to-Wasm wrapper in Turboshaft was already implemented in
wasm/wrappers.h and -inl.h.

- Lazy deoptimization is supported by reusing the builtin
`JSToWasmLazyDeoptContinuation` as in Turbofan.

Bug: 353475584
Change-Id: I23531af77f6ae078637ed5ce1fcf81bef6db1578
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6818818
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#102310}