[foozzie] Improve configuration of experimental fuzzing
This moves --experimental-fuzzing to the config experiments for
comparisons. Previously when it was in the default flags, it'd be
passed also to the baseline, which can lead to flag cycles, e.g.
if a flag is staged the implies optimizations, while the baseline
tries to switch off optimizations.

We run this with a probability of 50% like in other trials to also
retain coverage for paths where the staged flags are off.

We also bump the probability of --future to 50% to not reduce the
coverage of a flag when it promotes from --experimental-fuzzing to
--future.

Bug: 408971009
Change-Id: I07ef185f94de1f2389bba6aaff8494af4beec014
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6917698
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Auto-Submit: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102265}