[maglev] Remove {Load,Store}DoubleField
The LoadDoubleField instruction was misleadingly named. It didn't just
load a raw double value from an offset, but specifically loaded a tagged
HeapNumber from a field and then extracted its double value.

Likewise for StoreDoubleField.

This was confusing and could lead to incorrect usage, as the name
implied it could be used for any raw double field, which is not the
case.

This patch removes these instructions entirely. All usages are replaced
with the more explicit sequence of a LoadTaggedField (to get the
HeapNumber) followed by a {Load,Store}Float64. This makes the code
clearer and removes the potential for future misuse.

Change-Id: I9ee56f7890e63884682e564307eb8c94c1bc3fee
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6918957
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Auto-Submit: Jakob Linke <jgruber@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102266}