[test] Push golden files to Android devices when listing tests
Follow-up fix after:
https://crrev.com/c/6912786

After the change above, bytecode-expectation tests are set up
based on the golden files on disk. For that, the files need to
be present when the look-up happens.

Until now, we only pushed the required golden file to an
Android device when the respective test case was executed.
Now we already need all the golden files present, when the
test list is looked up.

The old individual pushing code is also still useful when the
test-runner executes only a single file with a filter.

Cq-Include-Trybots: luci.v8.try:v8_android_arm64_p7_rel
Change-Id: I6323ab0f4c36d76a37d4d4e4558fb4d6e39b2fe7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6920388
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102325}