[debug] Put Debugger.setScriptSource behind a feature flag
We plan to deprecate and remove the live edit functionality. This CL
puts the feature behind a flag that'll cause Debugger.setScriptSource
to fail with a ServerError in case the flag is disabled.

The CL also updates all affected tests to specifically enable the
feature.

Drive-by: Make the runtime function go through Debug::SetScriptSource
so also the tests in test/debugger honor the flag.

R=bmeurer@chromium.org

Bug: 434574729
Change-Id: Idd8b57ef67258c4946e66028b3df8427ce89b8da
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6929582
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102348}