[maglev] Improve non-eager inlining pipeline
Instead of running the optimizer for every function,
we run it after inlining all the known functions.

If we find a new function during the optimizer, we can
inline it and run the optimizer again.

Since finding a new function to inline during the optimizer
should be rarer (currently impossible). We also recalculate
phi use hints during the optimizer pass.

Bug: 424157317
Change-Id: Ibba9fc8902297c1138b0f09e57f838680dd59806
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6929209
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102357}