[turbolev] Recompute KNA
Recomputes the Known Node Aspects (KNA) for the entire graph.

KNA tracks information about nodes that can be used for
optimizations, such as eliminating redundant checks or loads.
Finding node alternatives, etc...

Currently, this is just an skeleton. We still need to register
things in the KNA, for instance save the maps for an object
after a CheckMaps. (follow up work).

This is done in a different processor and can be combine with
MaglevOptimizer to improve its efficiency.

Bug: 424157317
Change-Id: I5b568ee670fff950ff1e0806399b25bceae78cc8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6929962
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102381}