[heap] Avoid using ephemeron remembered set in full GC
The ephemeron_remembered_set() is used as the OLD_TO_NEW remembered
set for ephemeron keys. It is an optimization to avoid strongifying
such references in minor GCs.

After a full GC there are no OLD_TO_NEW references anymore, so we
are already able to drop this remembered set at the end of the GC.
But we still need this remembered set for updating pointers during the
GC for the same reason we need OLD_TO_NEW there as well: We don't
record pointers to new space in the OLD_TO_OLD remembered set. We
need it to find all pointers to new space objects for relocating them.

This CL is able to avoid usage of the ephemeron_remembered_set()
during the full GC entirely. During incremental marking the
write barrier and marker record pointers to young objects for
ephemeron keys in the regular OLD_TO_NEW remembered set instead.

With this the "update pointers" phase can treat ephemeron keys like
regular OLD_TO_NEW slots and doesn't need special handling anymore.

Change-Id: Id2ba6884e69ed01d8d09951b30060b10ea867937
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6917693
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102391}