[wasmfx][jspi] Allow JSPI to suspend multiple stacks at once
Fix all places where JSPI still assumes that it can only capture a
single stack. With the introduction of WasmFX, this assumption does
not hold anymore: we can switch stacks within a Suspender scope.

The main issue is the WasmSuspend builtin itself, which previously
jumped to the current stack's immediate parent. Now it needs to jump to
the stack that corresponds to the outer scope of the suspender. If we
are in nested WA.promising() calls, this is the stack of the parent
WasmSuspenderObject. But if we are already inside the outermost
WA.promising() call, the suspender does not have a parent.

To avoid a special case in the WasmSuspend builtin, this change
introduces a sentinel WasmSuspenderObject to represent the initial
stack. This way, the builtin can unconditionally jump to the parent
suspender's stack.

Additionally:
- We have to check for the presence of non-suspendable (JS/C) frames
  before suspending not just in the current stack, but also in the inner
  stacks,
- When resuming a chain of stacks, the current stack becomes the parent
  of the tail of the chain, not of the target stack.

R=clemensb@chromium.org

Bug: 388533754
Change-Id: I44c9b55598b309756dc4f5addd10636f1f89f190
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6929584
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102387}