[wasm][fuzzer] Store memory and global traces for differential fuzzing
This adds an option to store the memory and global traces instead of
printing them to stdout in the runtime functions. This is a preparation
for comparing the memory and global traces between the reference and
non-reference runs in the differential fuzzing, which requires the
traces to be stored.

Bug: 436219201
Change-Id: I6bbb268eaccc798c8336368fa2415fbc97d240b9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6897246
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Nikolas Kaipel <nikolaskaipel@google.com>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102417}