[riscv][wasm] Make non-far jump table entry patching atomic
Before this change, the writes would be atomic, but the reader would be
allowed to fetch and execute the two instructions separately. This meant
that it was possible to execute the unpatched 'auipc' instruction, but
the patched 'jalr' instruction - thus leading to jumps into the middle
of nowhere.

This fixes the flaky crashing in the wasm-spec-tests/return_call test.

Change-Id: Ic1b5da7add15760ea890ad0ef340c18e5bfe40e3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6943610
Commit-Queue: Kasper Lund <kasperl@rivosinc.com>
Reviewed-by: Florian Loitsch <floitsch@rivosinc.com>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102448}