[heap] Fix race in JSWeakRef handling
Current handling checked whether the target was already scavenged and
updated the pointer if it was. Then it would check again and add to the
worklist if it wasn't yet scavenged. This only works if no other
scavenger thread scavenges the target in between.
If another thread scavenged it in between, the target would not have
been updated nor was it added to the worklist to update it later,
resulting in UaF.

Fix by merging updating the slot and adding to the worklist to be under
one check instead of two.

Bug: 340777103
Change-Id: I0cc706397919b1db435cf91e2e2222630369fc34
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6943135
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102458}