[compiler][isel] Mark all processed nodes as Defined
Here is a pattern I observed in a Turboshaft graph:

  ...
  538: OverflowCheckedBinop
  539: Projection(#538)[0]
  540: Projection(#538)[1]
  542: DeoptimizeIf(#540)[Overflow]
  545: Store(..., #539)

During instruction selection, the Store is processed first, but is not
marked as Defined (because it doesn't produce a value), which means
that when processing DeoptimizeIf and calling
TryPrepareScheduleFirstProjection to define Projection[0], it fails
because Projection[0] has a use that's not already defined (the
Store). As a result, the generated instruction sequence does:

  (v57(0), v55(R)) = X64Add32 && set if overflow
  X64Cmp32 && deoptimize if not equal v55 #0 #10

thus materializing the overflow bit rather than fusing the add and the
deoptimize based on the overflow bit.

This CL fixes this by marking all visited nodes as Defined.

Change-Id: Ib65a13f884fdd507accb7d01812a61dc201c9f72
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6948072
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102504}